# Add quadtree.FindMatching(Point, Filter) method