@paulmach 2017-08-31
FYI, updated readme in https://github.com/paulmach/go.geo/commit/6f4efd27f1a7e05da0f7680e58421aabeb38798f

feel free to make other changes :)